chore: restructure XML reader/writer packages
