+++ plugins/com.cubrid.cubridmigration.core/src/com/cubrid/cubridmigration/core/engine/template/writer/node/SourceNodeWriter.java
@@ -27,7 +27,7 @@
  * OF SUCH DAMAGE.
  *
  */
-package com.cubrid.cubridmigration.core.engine.template.writer;
+package com.cubrid.cubridmigration.core.engine.template.writer.node;
 
 import static com.cubrid.cubridmigration.core.engine.template.MigrationTemplateUtils.*;
 import static com.cubrid.cubridmigration.core.engine.template.TemplateTags.*;
